backend/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.7
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
billiard==3.6.4.0
boto3==1.20.3
botocore==1.23.3
celery==5.1.2
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.2.1
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.1
djangorestframework-simplejwt
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.10
idna==2.10
importlib-metadata==3.1.1
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
msal==1.7.0
nodeenv==1.5.0
packaging==21.0
pre-commit==2.9.3
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.2.2
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
